+++ lib/sofabuffers.dart
@@ -35,7 +35,6 @@ export 'src/wire.dart'
         fixlenMax,
         arrayMax,
         maxDepth,
-        lazySeqDepth,
         WireType,
         FixlenType,
         DecodeStatus,